Lame Wireframe
This is the general wireframe we're working with. It gives a good mapping and idea of function and setup but we're still deciding a few specs as we go. It will certainly undergo A LOT of changes and updates as we proceed. I called it "Lame" because it should and will become prettier. 

Hannah and Ally