Lame Wireframe
This is the general wireframe we're working with. It serves as a good map and idea of our desired functions and set up but we will be making decisions on a few specs as we go. It is called "Lame" because it should and will be prettier. There will be a LOT of changes made to this.

Hannah and Ally